refactor: __init__.py -> snippets.py
